+++ src/Controller/Backoffice/MaincategoryController.php
@@ -19,6 +19,9 @@ class MaincategoryController extends AbstractController
 {
     /**
      * @Route("/", name="app_backoffice_maincategory_index", methods={"GET"})
+     *
+     * @param MainCategoryRepository $mainCategoryRepository
+     * @return Response
      */
     public function index(MainCategoryRepository $mainCategoryRepository): Response
     {
@@ -29,6 +32,11 @@ class MaincategoryController extends AbstractController
 
     /**
      * @Route("/new", name="app_backoffice_maincategory_new", methods={"GET", "POST"})
+     *
+     * @param Request $request
+     * @param MainCategoryRepository $mainCategoryRepository
+     * @param CustomSlugger $customSlugger
+     * @return Response
      */
     public function new(Request $request, MainCategoryRepository $mainCategoryRepository, CustomSlugger $customSlugger): Response
     {
@@ -57,9 +65,15 @@ class MaincategoryController extends AbstractController
 
     /**
      * @Route("/{id}", name="app_backoffice_maincategory_show", methods={"GET"})
+     *
+     * @param MainCategory|null $mainCategory
+     * @return Response
      */
-    public function show(MainCategory $mainCategory): Response
+    public function show(?MainCategory $mainCategory): Response
     {
+        if (!$mainCategory) {
+            throw $this->createNotFoundException("La main catégorie demandée n'a pas été trouvée");}
+
         return $this->render('backoffice/maincategory/show.html.twig', [
             'main_category' => $mainCategory,
         ]);
@@ -67,9 +81,18 @@ class MaincategoryController extends AbstractController
 
     /**
      * @Route("/{id}/edit", name="app_backoffice_maincategory_edit", methods={"GET", "POST"})
+     *
+     * @param Request $request
+     * @param MainCategory|null $mainCategory
+     * @param MainCategoryRepository $mainCategoryRepository
+     * @param CustomSlugger $customSlugger
+     * @return Response
      */
-    public function edit(Request $request, MainCategory $mainCategory, MainCategoryRepository $mainCategoryRepository, CustomSlugger $customSlugger): Response
+    public function edit(Request $request, ?MainCategory $mainCategory, MainCategoryRepository $mainCategoryRepository, CustomSlugger $customSlugger): Response
     {
+        if (!$mainCategory) {
+            throw $this->createNotFoundException("la main catégorie demandée n'a pas été trouvée");}
+
         $form = $this->createForm(MainCategoryType::class, $mainCategory);
         $form->handleRequest($request);
 
@@ -96,9 +119,17 @@ class MaincategoryController extends AbstractController
 
     /**
      * @Route("/{id}", name="app_backoffice_maincategory_delete", methods={"POST"})
+     *
+     * @param Request $request
+     * @param MainCategory|null $mainCategory
+     * @param MainCategoryRepository $mainCategoryRepository
+     * @return Response
      */
-    public function delete(Request $request, MainCategory $mainCategory, MainCategoryRepository $mainCategoryRepository): Response
+    public function delete(Request $request, ?MainCategory $mainCategory, MainCategoryRepository $mainCategoryRepository): Response
     {
+        if (!$mainCategory) {
+            throw $this->createNotFoundException("La main catégorie demandée n'a pas été trouvée");}
+
         if ($this->isCsrfTokenValid('delete' . $mainCategory->getId(), $request->request->get('_token'))) {
             $mainCategoryRepository->remove($mainCategory, true);
