+++ src/Controller/Backoffice/OfferController.php
@@ -4,7 +4,6 @@ namespace App\Controller\Backoffice;
 
 use App\Entity\Offer;
 use App\Form\OfferType;
-use App\Form\OfferTypeCustom;
 use App\Repository\OfferRepository;
 use DateTime;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -19,9 +18,15 @@ class OfferController extends AbstractController
 {
     /**
      * @Route("/{id}", name="app_backoffice_offer_show", methods={"GET"})
+     *
+     * @param Offer|null $offer
+     * @return Response
      */
-    public function show(Offer $offer): Response
+    public function show(?Offer $offer): Response
     {
+        if (!$offer) {
+            throw $this->createNotFoundException("L'offre demandée n'a pas été trouvée");}
+
         return $this->render('backoffice/offer/show.html.twig', [
             'offer' => $offer,
         ]);
@@ -29,9 +34,17 @@ class OfferController extends AbstractController
 
     /**
      * @Route("/{id}/edit", name="app_backoffice_offer_edit", methods={"GET", "POST"})
+     *
+     * @param Request $request
+     * @param Offer|null $offer
+     * @param OfferRepository $offerRepository
+     * @return Response
      */
-    public function edit(Request $request, Offer $offer, OfferRepository $offerRepository): Response
+    public function edit(Request $request, ?Offer $offer, OfferRepository $offerRepository): Response
     {
+        if (!$offer) {
+            throw $this->createNotFoundException("L'offre demandée n'a pas été trouvée");}
+
         $form = $this->createForm(OfferType::class, $offer);
         $form->handleRequest($request);
 
@@ -54,9 +67,17 @@ class OfferController extends AbstractController
 
     /**
      * @Route("/{id}", name="app_backoffice_offer_delete", methods={"POST"})
+     *
+     * @param Request $request
+     * @param Offer|null $offer
+     * @param OfferRepository $offerRepository
+     * @return Response
      */
-    public function delete(Request $request, Offer $offer, OfferRepository $offerRepository): Response
+    public function delete(Request $request, ?Offer $offer, OfferRepository $offerRepository): Response
     {
+        if (!$offer) {
+            throw $this->createNotFoundException("L'offre demandée n'a pas été trouvée");}
+
         if ($this->isCsrfTokenValid('delete' . $offer->getId(), $request->request->get('_token'))) {
             $offerRepository->remove($offer, true);
 
@@ -68,17 +89,23 @@ class OfferController extends AbstractController
 
     /**
      * @Route("/{id}/validate", name="app_backoffice_offer_validate", methods={"POST"})
+     *
+     * @param Request $request
+     * @param Offer|null $offer
+     * @param OfferRepository $offerRepository
+     * @return Response
      */
-    public function validate(Request $request, Offer $offer, OfferRepository $offerRepository): Response
+    public function validate(Request $request, ?Offer $offer, OfferRepository $offerRepository): Response
     {
-        // if ($this->isCsrfTokenValid('validate' . $offer->getId(), $request->request->get('_token'))) {
-        // }
-        
+        if ($this->isCsrfTokenValid('validate' . $offer->getId(), $request->request->get('_token'))) {
+        }
+        if (!$offer) {
+            throw $this->createNotFoundException("L'offre demandée n'a pas été trouvée");}
         $offer->setIsReported(false);
         $offer->setUpdatedAt(new DateTime());
         $offerRepository->add($offer, true);
 
 
-        return $this->redirectToRoute('app_backoffice_reported', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_backoffice_reported_index', [], Response::HTTP_SEE_OTHER);
     }
 }
